refactor: move execution to main.py
